refactor: Refactor Banner Group Function(#11)
