+++ src/components/Banner/index.js
@@ -1,0 +1,3 @@
+import Banner from './Banner';
+
+export { Banner };
