+++ src/components/admin/Marketing/components/ConfigurableBanner.jsx
@@ -1,15 +1,16 @@
-import Banner from '@/components/Banner';
+import { useEffect, useMemo } from 'react';
+
+import { Banner } from '@/components/Banner';
 import { useSizeType } from '@/context/sizeTypeContext';
 import backend from '@/service/backend';
 import { getBannerByType } from '@/service/banner/banner-utils';
-import { useEffect } from 'react';
-import { useMemo } from 'react';
+
 import { useAdminBannerState } from '../atoms/adminBanner.atom';
 import ConfigurableMainLayout from './ConfigurableMainLayout';
 import ConfigurableOption from './ConfigurableMainOption';
 
 export default function ConfigurableBanner() {
-  // ! use RecoilState for shareing Sidebar Menu Setting
+  // ! use RecoilState for sharing Sidebar Menu Setting
   const [banners, setBanners] = useAdminBannerState();
   const sizeType = useSizeType();
