+++ src/components/admin/Marketing/components/ConfigurationSequenceOptionMenu.jsx
@@ -3,32 +3,23 @@ import { useAdminBannerState } from '../atoms/adminBanner.atom';
 import { BannerSequenceForm } from './BannerSequenceForm';
 
 import {
-  Container,
   Header,
   HeaderTitle,
   HeaderButton,
   WrapForm,
 } from './ConfigurationSequenceOptionMenu.style';
 
-import {
-  filterMainBanner,
-  groupBannerToPairItem,
-} from '@/service/banner/banner-utils';
+import { filterMainBanner } from '@/service/banner/banner-utils';
 
 import { sequenceFromPairedBanner } from './BannerSequenceForm/utils';
-import { groupBy } from '@/utils';
 import backend from '@/service/backend';
 import { updateSequencesRequest } from '@/service/banner/banner-request';
 
 export default function ConfigurationSequenceOptionMenu() {
-  const [rawBanners] = useAdminBannerState();
+  const [banners] = useAdminBannerState();
 
   const [sequences, setSequences] = useState(
-    sequenceFromPairedBanner(
-      filterMainBanner(
-        groupBannerToPairItem(groupBy(rawBanners, 'redirectUrl')),
-      ),
-    ) || [],
+    sequenceFromPairedBanner(filterMainBanner(banners)) || [],
   );
 
   function handleSubmit() {
@@ -36,7 +27,7 @@ export default function ConfigurationSequenceOptionMenu() {
   }
 
   return (
-    <Container>
+    <>
       <Header>
         <HeaderTitle>배너목록</HeaderTitle>
         <HeaderButton onClick={handleSubmit}>적용</HeaderButton>
@@ -44,6 +35,6 @@ export default function ConfigurationSequenceOptionMenu() {
       <WrapForm>
         <BannerSequenceForm setSequence={setSequences} sequence={sequences} />
       </WrapForm>
-    </Container>
+    </>
   );
 }
