+++ src/components/admin/Marketing/components/ConfigurationSequenceOptionMenu.style.js
@@ -1,9 +1,5 @@
 import styled from 'styled-components';
 
-export const Container = styled.div`
-  padding: 24px 32px;
-`;
-
 export const Header = styled.div`
   display: flex;
   justify-content: space-between;
